+++ bash/functions.sh
@@ -167,3 +167,195 @@ function get_config() {
     fi
     echo $setting
 }
+
+
+## ASSIGN DEVICES TO DECODERS
+
+function ask_for_device_assignments() {
+    log_heading "Gather information required to configure the decoder(s)"
+
+    decoder_being_installed=$1
+    decoder_count=1
+
+    log_message "Checking if an ACARS decoder is installed"
+    acars_decoder_installed="false"
+    if [[ -f /usr/local/bin/acarsdec ]]; then
+        log_message "The ACARSDEC decoder appears to be installed"
+        acars_decoder_installed="true"
+        acars_decoder="acarsdec"
+    fi
+    if [[ "${acars_decoder_installed}" == "true" && "${acars_decoder}" != "${decoder_being_installed}" ]]; then
+        decoder_count=$((decoder_count+1))
+    fi
+
+    log_message "Checking if an ADS-B decoder is installed"
+    adsb_decoder_installed="false"
+    if [[ $(dpkg-query -W -f='${STATUS}' dump1090-fa 2>/dev/null | grep -c "ok installed") -eq 1 ]]; then
+        log_message "The FlightAware dump1090 decoder appears to be installed"
+        adsb_decoder_installed="true"
+        adsb_decoder="dump1090-fa"
+    fi
+    if [[ "${adsb_decoder_installed}" == "true" && "${adsb_decoder}" != "${decoder_being_installed}" ]]; then
+        decoder_count=$((decoder_count+1))
+    fi
+
+    log_message "Checking if a UAT decoder is installed"
+    uat_decoder_installed="false"
+    if [[ $(dpkg-query -W -f='${STATUS}' dump978-fa 2>/dev/null | grep -c "ok installed") -eq 1 ]]; then
+        log_message "The FlightAware dump978 decoder appears to be installed"
+        uat_decoder_installed="true"
+        uat_decoder="dump978-fa"
+    fi
+    if [[ "${uat_decoder_installed}" == "true" && "${uat_decoder}" != "${decoder_being_installed}" ]]; then
+        decoder_count=$((decoder_count+1))
+    fi
+
+    log_message "Checking if a VDL Mode 2 decoder is installed"
+    vdlm2_decoder_installed="false"
+    if [[ -f /usr/local/bin/dumpvdl2 ]]; then
+        log_message "The dumpvdl2 decoder appears to be installed"
+        vdlm2_decoder_installed="true"
+        vdlm2_decoder="dumpvdl2"
+    fi
+    if [[ -f /usr/local/bin/vdlm2dec ]]; then
+        log_message "The VDLM2DEC decoder appears to be installed"
+        vdlm2_decoder_installed="true"
+        vdlm2_decoder="vdlm2dec"
+    fi
+    if [[ "${vdlm2_decoder_installed}" == "true" && "${vdlm2_decoder}" != "${decoder_being_installed}" ]]; then
+        decoder_count=$((decoder_count+1))
+    fi
+
+    if [[ $decoder_count > 1 ]]; then
+        log_message "Informing the user that existing decoder(s) appears to be installed"
+        whiptail --backtitle "Decoder Configuration" \
+                --title "RTL-SDR Dongle Assignments" \
+                --msgbox "It appears that existing decoder(s) have been installed on this device. In order to run this decoder in tandem with other decoders you will need to specifiy which RTL-SDR dongle each decoder is to use.\n\nKeep in mind in order to run multiple decoders on a single device you will need to have multiple RTL-SDR devices connected to your device." \
+                12 78
+
+        if [[ "${acars_decoder_installed}" == "true" && "${acars_decoder}" == "acarsdec" ]]; then
+            log_message "Determining which device is currently assigned to the ACARS decoder"
+            exec_start=`get_config "ExecStart" "/etc/systemd/system/acarsdec.service"`
+            RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER=`echo $exec_start | grep -o -P '(?<=-r ).*(?= -A)'`
+            log_message "Asking the user to assign a RTL-SDR device number to ACARSDEC"
+            acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number"
+            while [[ -z $acars_device_number ]]; do
+                acars_device_number=$(whiptail --backtitle "Decoder Configuration" \
+                                               --title "${acars_device_number_title}" \
+                                               --inputbox "\nEnter the RTL-SDR device number to assign your ACARSDEC decoder." \
+                                               8 78 \
+                                               "${RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER}" 3>&1 1>&2 2>&3)
+                exit_status=$?
+                if [[ $exit_status != 0 ]]; then
+                    exit 1
+                fi
+                acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number (REQUIRED)"
+            done
+        fi
+
+        if [[ "${adsb_decoder_installed}" == "true" && "${adsb_decoder}" == "dump1090-fa" ]]; then
+            log_message "Determining which device is currently assigned to the dump1090-fa decoder"
+            RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER=`get_config "RECEIVER_SERIAL" "/etc/default/dump1090-fa"`
+            log_message "Asking the user to assign a RTL-SDR device number to dump1090-fa"
+            adsb_device_number_title="Enter the dump1090-fa RTL-SDR Device Number"
+            while [[ -z $adsb_device_number ]]; do
+                adsb_device_number=$(whiptail --backtitle "Decoder Configuration" \
+                                            --title "${adsb_device_number_title}" \
+                                            --inputbox "\nEnter the RTL-SDR device number to assign to dump1090-fa." \
+                                            8 78 \
+                                            "${RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER}" 3>&1 1>&2 2>&3)
+                exit_status=$?
+                if [[ $exit_status != 0 ]]; then
+                    exit 1
+                fi
+                adsb_device_number_title="Enter the dump1090-fa RTL-SDR Device Number (REQUIRED)"
+            done
+        fi
+
+        if [[ "${uat_decoder_installed}" == "true" && "${dump978-fa}" == "dump978-fa" ]]; then
+            log_message "Determining which device is currently assigned to dump978-fa"
+            receiver_options=`get_config "RECEIVER_OPTIONS" "/etc/default/dump978-fa"`
+            RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER=$receiver_options | grep -o -P '(?<=serial=).*(?= --)'
+            log_message "Asking the user to assign a RTL-SDR device number to dump978-fa"
+            uat_device_number_title="Enter the dump978-fa RTL-SDR Device Number"
+            while [[ -z $uat_device_number ]] ; do
+                uat_device_number=$(whiptail --backtitle "Decoder Configuration" \
+                                            --title "${uat_device_number_title}" \
+                                            --inputbox "\nEnter the RTL-SDR device number to assign to dump978-fa." \
+                                            8 78 \
+                                            "${RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER}" 3>&1 1>&2 2>&3)
+                exit_status=$?
+                if [[ $exit_status != 0 ]]; then
+                    exit 1
+                fi
+                uat_device_number_title="Enter the dump978-fa RTL-SDR Device Number (REQUIRED)"
+            done
+        fi
+
+        if [[ "${vdlm2_decoder_installed}" == "true" && "${vdlm2_decoder}" == "dumpvdl2" ]]; then
+            log_message "Determining which device is currently assigned to dumpvdl2"
+            exec_start=`get_config "ExecStart" "/etc/systemd/system/dumpvdl2.service"`
+            RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER=`echo $exec_start | grep -o -P '(?<=--rtlsdr ).*(?= --gain)'`
+            log_message "Asking the user to assign a RTL-SDR device number to dumpvdl2"
+            vdlm2_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number"
+            while [[ -z $vdlm2_device_number ]]; do
+                vdlm2_device_number=$(whiptail --backtitle "Decoder Configuration" \
+                                            --title "${vdlm2_device_number_title}" \
+                                            --inputbox "Enter the RTL-SDR device number to assign to dumpvdl2." \
+                                            8 78 \
+                                            "${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER}" 3>&1 1>&2 2>&3)
+                exit_status=$?
+                if [[ $exit_status != 0 ]]; then
+                    exit 1
+                fi
+                vdlm2_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number (REQUIRED)"
+            done
+        fi
+    fi
+}
+
+function assign_devices_to_decoders() {
+
+    log_heading "Configure decoders if more than one is present"
+
+    if [[ ! -z $RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER && "${RECEIVER_ACARS_DECODER_SOFTWARE}" == "acarsdec" ]]; then
+        log_message "Assigning RTL-SDR device number ${RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER} to ACARSDEC"
+        sudo sed -i -e "s|\(.*-r \)\([0-9]\+\)\( .*\)|\1${RECEIVER_DEVICE_ASSIGNED_TO_ACARS_DECODER}\3|g" /etc/systemd/system/acarsdec.service
+        log_message "Reload systemd units"
+        sudo systemctl daemon-reload
+        log_message "Restarting ACARSDEC"
+        sudo systemctl restart acarsdec
+    fi
+
+    if [[ ! -z $RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER && "${RECEIVER_ADSB_DECODER_SOFTWARE}" == "dump1090-fa" ]]; then
+        log_message "Assigning RTL-SDR device number ${RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER} to the FlightAware Dump1090 decoder"
+        change_config "RECEIVER_SERIAL" $RECEIVER_DEVICE_ASSIGNED_TO_ADSB_DECODER "/etc/default/dump1090-fa"
+        log_message "Restarting dump1090-fa"
+        sudo systemctl restart dump1090-fa
+    fi
+
+    if [[ ! -z $RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER && "${RECEIVER_UAT_DECODER_SOFTWARE}" == "dump978-fa" ]]; then
+        log_message "Assigning RTL-SDR device number ${RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER} to the FlightAware Dump978 decoder"
+        sudo sed -i -e "s|driver=rtlsdr|driver=rtlsdr,serial=${RECEIVER_DEVICE_ASSIGNED_TO_UAT_DECODER}|g" /etc/default/dump978-fa
+        log_message "Restarting dump978-fa"
+        sudo systemctl restart dump978-fa
+    fi
+
+    if [[ ! -z $RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER && "${RECEIVER_VDLM2_DECODER_SOFTWARE}" == "dumpvdl2" ]]; then
+        log_message "Assigning RTL-SDR device number ${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER} to dumpvdl2"
+        sudo sed -i -e "s|\(.*--rtlsdr \)\([0-9]\+\)\( .*\)|\1${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER}\3|g" /etc/systemd/system/dumpvdl2.service
+        log_message "Reloading systemd units"
+        sudo systemctl daemon-reload
+        log_message "Restarting dumpvdl2"
+        sudo systemctl restart dumpvdl2
+    fi
+
+    if [[ ! -z "${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER}" && "${RECEIVER_VDLM2_DECODER_SOFTWARE}" == "vdlm2dec" ]]; then
+        log_message "Assigning RTL-SDR device number ${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER} to vdlm2dec"
+        sudo sed -i -e "s|\(.*-r \)\([0-9]\+\)\( .*\)|\1${RECEIVER_DEVICE_ASSIGNED_TO_VDLM2_DECODER}\3|g" /etc/systemd/system/vdlm2dec.service
+        log_message "Reloading systemd units"
+        sudo systemctl daemon-reload
+        log_message "Restarting vdlm2dec"
+        sudo systemctl restart vdlm2dec
+    fi
+}